refactor: share AOT services from core
