+++ micronaut-maven-aot-core/src/main/java/io/micronaut/maven/aot/AbstractAotAnalysisMojo.java
@@ -16,9 +16,9 @@
 package io.micronaut.maven.aot;
 
 import io.micronaut.aot.std.sourcegen.KnownMissingTypesSourceGenerator;
-import io.micronaut.maven.aot.internal.AotCompilerService;
-import io.micronaut.maven.aot.internal.AotDependencyResolutionService;
-import io.micronaut.maven.aot.internal.AotExecutorService;
+import io.micronaut.maven.services.CompilerService;
+import io.micronaut.maven.services.DependencyResolutionService;
+import io.micronaut.maven.services.ExecutorService;
 import org.apache.commons.io.FileUtils;
 import org.apache.maven.execution.MavenSession;
 import org.apache.maven.plugin.MojoExecutionException;
@@ -62,10 +62,10 @@ public abstract class AbstractAotAnalysisMojo extends AbstractMicronautAotCliMoj
     private File configFile;
 
     @Inject
-    protected AbstractAotAnalysisMojo(AotCompilerService compilerService,
-                                      AotExecutorService executorService,
+    protected AbstractAotAnalysisMojo(CompilerService compilerService,
+                                      ExecutorService executorService,
                                       MavenProject mavenProject,
-                                      AotDependencyResolutionService dependencyResolutionService,
+                                      DependencyResolutionService dependencyResolutionService,
                                       MavenSession mavenSession,
                                       ToolchainManager toolchainManager) {
         super(compilerService, executorService, mavenProject, dependencyResolutionService, mavenSession, toolchainManager);
