--- micronaut-maven-aot-core/src/main/java/io/micronaut/maven/aot/internal/AotDependencyResolutionService.java
@@ -1,115 +1,0 @@
-/*
- * Copyright 2017-2022 original authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.micronaut.maven.aot.internal;
-
-import org.apache.maven.execution.MavenSession;
-import org.apache.maven.model.DependencyManagement;
-import org.apache.maven.project.MavenProject;
-import org.eclipse.aether.RepositorySystem;
-import org.eclipse.aether.RepositorySystemSession;
-import org.eclipse.aether.artifact.Artifact;
-import org.eclipse.aether.artifact.DefaultArtifact;
-import org.eclipse.aether.collection.CollectRequest;
-import org.eclipse.aether.graph.Dependency;
-import org.eclipse.aether.graph.DependencyFilter;
-import org.eclipse.aether.resolution.ArtifactResult;
-import org.eclipse.aether.resolution.DependencyRequest;
-import org.eclipse.aether.resolution.DependencyResolutionException;
-import org.eclipse.aether.resolution.DependencyResult;
-import org.eclipse.aether.util.artifact.JavaScopes;
-import org.eclipse.aether.util.filter.DependencyFilterUtils;
-
-import javax.inject.Inject;
-import javax.inject.Singleton;
-import java.io.File;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.stream.Stream;
-
-/**
- * Dependency resolution helpers used by the extracted AOT mojos.
- */
-@Singleton
-public final class AotDependencyResolutionService {
-
-    private final MavenSession mavenSession;
-    private final MavenProject mavenProject;
-    private final RepositorySystem repositorySystem;
-
-    @Inject
-    public AotDependencyResolutionService(MavenSession mavenSession,
-                                          MavenProject mavenProject,
-                                          RepositorySystem repositorySystem) {
-        this.mavenSession = mavenSession;
-        this.mavenProject = mavenProject;
-        this.repositorySystem = repositorySystem;
-    }
-
-    public static List<String> toClasspath(List<ArtifactResult> resolutionResult) {
-        return resolutionResult.stream()
-            .map(ArtifactResult::getArtifact)
-            .map(Artifact::getFile)
-            .map(File::getAbsolutePath)
-            .toList();
-    }
-
-    public List<ArtifactResult> artifactResultsFor(Stream<Artifact> artifacts, boolean applyManagedDependencies) throws DependencyResolutionException {
-        RepositorySystemSession repositorySession = mavenSession.getRepositorySession();
-        DependencyFilter classpathFilter = DependencyFilterUtils.classpathFilter(JavaScopes.RUNTIME);
-        CollectRequest collectRequest = new CollectRequest();
-        collectRequest.setRepositories(mavenProject.getRemoteProjectRepositories());
-
-        if (applyManagedDependencies) {
-            DependencyManagement dependencyManagement = mavenProject.getDependencyManagement();
-            List<org.apache.maven.model.Dependency> dependencies = dependencyManagement == null || dependencyManagement.getDependencies() == null
-                ? List.of()
-                : dependencyManagement.getDependencies();
-            HashMap<String, Dependency> dependencyMap = HashMap.newHashMap(dependencies.size());
-            for (org.apache.maven.model.Dependency dependency : dependencies) {
-                String ga = dependency.getGroupId() + ":" + dependency.getArtifactId();
-                Artifact artifact = new DefaultArtifact(
-                    dependency.getGroupId(),
-                    dependency.getArtifactId(),
-                    dependency.getClassifier(),
-                    dependency.getType(),
-                    dependency.getVersion()
-                );
-                dependencyMap.putIfAbsent(ga, new Dependency(artifact, dependency.getScope(), Boolean.valueOf(dependency.getOptional())));
-            }
-            collectRequest.setManagedDependencies(new ArrayList<>(dependencyMap.values()));
-            artifacts.forEach(artifact -> {
-                if (artifact.getVersion() == null || artifact.getVersion().isEmpty()) {
-                    dependencyMap.computeIfPresent(artifact.getGroupId() + ":" + artifact.getArtifactId(), (coordinates, dependency) -> {
-                        collectRequest.addDependency(new Dependency(
-                            new DefaultArtifact(artifact.getGroupId(), artifact.getArtifactId(), artifact.getExtension(), dependency.getArtifact().getVersion()),
-                            JavaScopes.RUNTIME
-                        ));
-                        return dependency;
-                    });
-                } else {
-                    collectRequest.addDependency(new Dependency(artifact, JavaScopes.RUNTIME));
-                }
-            });
-        } else {
-            artifacts.map(artifact -> new Dependency(artifact, JavaScopes.RUNTIME)).forEach(collectRequest::addDependency);
-        }
-
-        DependencyRequest dependencyRequest = new DependencyRequest(collectRequest, classpathFilter);
-        DependencyResult dependencyResult = repositorySystem.resolveDependencies(repositorySession, dependencyRequest);
-        return dependencyResult.getArtifactResults();
-    }
-}
