+++ micronaut-maven-aot-plugin/src/main/java/io/micronaut/maven/aot/AotSampleMojo.java
@@ -15,9 +15,9 @@
  */
 package io.micronaut.maven.aot;
 
-import io.micronaut.maven.aot.internal.AotCompilerService;
-import io.micronaut.maven.aot.internal.AotDependencyResolutionService;
-import io.micronaut.maven.aot.internal.AotExecutorService;
+import io.micronaut.maven.services.CompilerService;
+import io.micronaut.maven.services.DependencyResolutionService;
+import io.micronaut.maven.services.ExecutorService;
 import org.apache.maven.execution.MavenSession;
 import org.apache.maven.plugins.annotations.LifecyclePhase;
 import org.apache.maven.plugins.annotations.Mojo;
@@ -34,10 +34,10 @@ import javax.inject.Inject;
 public class AotSampleMojo extends AbstractAotSampleMojo {
 
     @Inject
-    public AotSampleMojo(AotCompilerService compilerService,
-                         AotExecutorService executorService,
+    public AotSampleMojo(CompilerService compilerService,
+                         ExecutorService executorService,
                          MavenProject mavenProject,
-                         AotDependencyResolutionService dependencyResolutionService,
+                         DependencyResolutionService dependencyResolutionService,
                          MavenSession mavenSession,
                          ToolchainManager toolchainManager) {
         super(compilerService, executorService, mavenProject, dependencyResolutionService, mavenSession, toolchainManager);
