+++ micronaut-maven-plugin/src/main/java/io/micronaut/maven/aot/AotAnalysisMojo.java
@@ -15,9 +15,9 @@
  */
 package io.micronaut.maven.aot;
 
-import io.micronaut.maven.aot.internal.AotCompilerService;
-import io.micronaut.maven.aot.internal.AotDependencyResolutionService;
-import io.micronaut.maven.aot.internal.AotExecutorService;
+import io.micronaut.maven.services.CompilerService;
+import io.micronaut.maven.services.DependencyResolutionService;
+import io.micronaut.maven.services.ExecutorService;
 import org.apache.maven.execution.MavenSession;
 import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.ResolutionScope;
@@ -42,8 +42,8 @@ public class AotAnalysisMojo extends AbstractAotAnalysisMojo {
 
     @Inject
     @SuppressWarnings("CdiInjectionPointsInspection")
-    public AotAnalysisMojo(AotCompilerService compilerService, AotExecutorService executorService, MavenProject mavenProject,
-                           AotDependencyResolutionService dependencyResolutionService,
+    public AotAnalysisMojo(CompilerService compilerService, ExecutorService executorService, MavenProject mavenProject,
+                           DependencyResolutionService dependencyResolutionService,
                            MavenSession mavenSession, ToolchainManager toolchainManager) {
         super(compilerService, executorService, mavenProject, dependencyResolutionService, mavenSession, toolchainManager);
     }
